+++ tests/BUILD
@@ -39,6 +39,12 @@ jax_multiplatform_test(
 )
 
 jax_multiplatform_test(
+    name = "debug_info_test",
+    srcs = ["debug_info_test.py"],
+    enable_configs = ["tpu_v3_2x2"],
+)
+
+jax_multiplatform_test(
     name = "device_test",
     srcs = ["device_test.py"],
 )
